Collect whenthen tests in one module
