+++ py-polars/tests/unit/datatypes/test_object.py
@@ -5,22 +5,6 @@ import numpy as np
 import polars as pl
 
 
-def test_object_when_then_4702() -> None:
-    # please don't ever do this
-    x = pl.DataFrame({"Row": [1, 2], "Type": [pl.Date, pl.UInt8]})
-
-    assert x.with_columns(
-        pl.when(pl.col("Row") == 1)
-        .then(pl.lit(pl.UInt16, allow_object=True))
-        .otherwise(pl.lit(pl.UInt8, allow_object=True))
-        .alias("New_Type")
-    ).to_dict(False) == {
-        "Row": [1, 2],
-        "Type": [pl.Date, pl.UInt8],
-        "New_Type": [pl.UInt16, pl.UInt8],
-    }
-
-
 def test_object_empty_filter_5911() -> None:
     df = pl.DataFrame(
         data=[
