+++ py-polars/tests/unit/functions/test_whenthen.py
@@ -1,0 +1,159 @@
+from datetime import datetime
+
+import polars as pl
+from polars.testing import assert_frame_equal
+
+
+def test_when_then_implicit_none() -> None:
+    df = pl.DataFrame(
+        {
+            "team": ["A", "A", "A", "B", "B", "C"],
+            "points": [11, 8, 10, 6, 6, 5],
+        }
+    )
+
+    assert df.select(
+        pl.when(pl.col("points") > 7).then("Foo"),
+        pl.when(pl.col("points") > 7).then("Foo").alias("bar"),
+    ).to_dict(False) == {
+        "literal": ["Foo", "Foo", "Foo", None, None, None],
+        "bar": ["Foo", "Foo", "Foo", None, None, None],
+    }
+
+
+def test_when_then_empty_list_5547() -> None:
+    out = pl.DataFrame({"a": []}).select([pl.when(pl.col("a") > 1).then([1])])
+    assert out.shape == (0, 1)
+    assert out.dtypes == [pl.List(pl.Int64)]
+
+
+def test_nested_when_then_and_wildcard_expansion_6284() -> None:
+    df = pl.DataFrame(
+        {
+            "1": ["a", "b"],
+            "2": ["c", "d"],
+        }
+    )
+
+    out0 = df.with_columns(
+        pl.when(pl.any_horizontal(pl.all() == "a"))
+        .then("a")
+        .otherwise(
+            pl.when(pl.any_horizontal(pl.all() == "d")).then("d").otherwise(None)
+        )
+        .alias("result")
+    )
+
+    out1 = df.with_columns(
+        pl.when(pl.any_horizontal(pl.all() == "a"))
+        .then("a")
+        .when(pl.any_horizontal(pl.all() == "d"))
+        .then("d")
+        .otherwise(None)
+        .alias("result")
+    )
+
+    assert_frame_equal(out0, out1)
+    assert out0.to_dict(False) == {
+        "1": ["a", "b"],
+        "2": ["c", "d"],
+        "result": ["a", "d"],
+    }
+
+
+def test_list_zip_with_logical_type() -> None:
+    df = pl.DataFrame(
+        {
+            "start": [datetime(2023, 1, 1, 1, 1, 1), datetime(2023, 1, 1, 1, 1, 1)],
+            "stop": [datetime(2023, 1, 1, 1, 3, 1), datetime(2023, 1, 1, 1, 4, 1)],
+            "use": [1, 0],
+        }
+    )
+
+    df = df.with_columns(
+        pl.date_ranges(
+            pl.col("start"), pl.col("stop"), interval="1h", eager=False, closed="left"
+        ).alias("interval_1"),
+        pl.date_ranges(
+            pl.col("start"), pl.col("stop"), interval="1h", eager=False, closed="left"
+        ).alias("interval_2"),
+    )
+
+    out = df.select(
+        pl.when(pl.col("use") == 1)
+        .then(pl.col("interval_2"))
+        .otherwise(pl.col("interval_1"))
+        .alias("interval_new")
+    )
+    assert out.dtypes == [pl.List(pl.Datetime(time_unit="us", time_zone=None))]
+
+
+def test_type_coercion_when_then_otherwise_2806() -> None:
+    out = (
+        pl.DataFrame({"names": ["foo", "spam", "spam"], "nrs": [1, 2, 3]})
+        .select(
+            [
+                pl.when(pl.col("names") == "spam")
+                .then(pl.col("nrs") * 2)
+                .otherwise(pl.lit("other"))
+                .alias("new_col"),
+            ]
+        )
+        .to_series()
+    )
+    expected = pl.Series("new_col", ["other", "4", "6"])
+    assert out.to_list() == expected.to_list()
+
+    # test it remains float32
+    assert (
+        pl.Series("a", [1.0, 2.0, 3.0], dtype=pl.Float32)
+        .to_frame()
+        .select(pl.when(pl.col("a") > 2.0).then(pl.col("a")).otherwise(0.0))
+    ).to_series().dtype == pl.Float32
+
+
+def test_when_then_edge_cases_3994() -> None:
+    df = pl.DataFrame(data={"id": [1, 1], "type": [2, 2]})
+
+    # this tests if lazy correctly assigns the list schema to the column aggregation
+    assert (
+        df.lazy()
+        .groupby(["id"])
+        .agg(pl.col("type"))
+        .with_columns(
+            pl.when(pl.col("type").list.lengths() == 0)
+            .then(pl.lit(None))
+            .otherwise(pl.col("type"))
+            .keep_name()
+        )
+        .collect()
+    ).to_dict(False) == {"id": [1], "type": [[2, 2]]}
+
+    # this tests ternary with an empty argument
+    assert (
+        df.filter(pl.col("id") == 42)
+        .groupby(["id"])
+        .agg(pl.col("type"))
+        .with_columns(
+            pl.when(pl.col("type").list.lengths() == 0)
+            .then(pl.lit(None))
+            .otherwise(pl.col("type"))
+            .keep_name()
+        )
+    ).to_dict(False) == {"id": [], "type": []}
+
+
+def test_object_when_then_4702() -> None:
+    # please don't ever do this
+    x = pl.DataFrame({"Row": [1, 2], "Type": [pl.Date, pl.UInt8]})
+
+    assert x.with_columns(
+        pl.when(pl.col("Row") == 1)
+        .then(pl.lit(pl.UInt16, allow_object=True))
+        .otherwise(pl.lit(pl.UInt8, allow_object=True))
+        .alias("New_Type")
+    ).to_dict(False) == {
+        "Row": [1, 2],
+        "Type": [pl.Date, pl.UInt8],
+        "New_Type": [pl.UInt16, pl.UInt8],
+    }
